database migration with checks

+++ database/migrations/2016_06_21_151815_create_failed_jobs_table.php
@@ -12,13 +12,15 @@ class CreateFailedJobsTable extends Migration
      */
     public function up()
     {
-        Schema::create('failed_jobs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->text('connection');
-            $table->text('queue');
-            $table->longText('payload');
-            $table->timestamp('failed_at')->useCurrent();
-        });
+        if (!Schema::hasTable('failed_jobs')) {
+            Schema::create('failed_jobs', function (Blueprint $table) {
+                $table->increments('id');
+                $table->text('connection');
+                $table->text('queue');
+                $table->longText('payload');
+                $table->timestamp('failed_at')->useCurrent();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,8 @@ class CreateFailedJobsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('failed_jobs');
+        if (Schema::hasTable('failed_jobs')) {
+            Schema::drop('failed_jobs');
+        }
     }
 }
